backend/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication (JWT)
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# PostgreSQL + PostGIS
psycopg2-binary==2.9.10
django-environ==0.11.2

# Geospatial
GDAL==3.10.1

# Django Channels (WebSockets)
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Celery (Async Task Queue)
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# AI / NLP
openai==1.58.1
tiktoken==0.8.0

# Utilities
Pillow==11.1.0
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1

# Code Quality
flake8==7.1.1
black==24.10.0
isort==5.13.2
